TANKX/collect_ticker_data/requirements.txt
redis
python-binance
python-dotenv